django-bakend/requirements-dev.txt
-r requirements.txt
pytest==8.3.4
pytest-django==4.9.0
factory_boy==3.3.1
